pyproject.toml
[project]
name = "azure-convention-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-search-documents>=11.5.3",
    "dotenv>=0.9.9",
    "openai>=1.107.3",
    "pandas>=2.3.2",
    "streamlit>=1.49.1",
]
